+++ src/main/java/cz/cvut/fel/aic/agentpolis/simmodel/entity/MovingEntity.java
@@ -27,7 +27,7 @@ import cz.cvut.fel.aic.agentpolis.simmodel.environment.transportnetwork.elements
  * @author fido
  */
 public interface MovingEntity {
-	public double getVelocity();
+	public int getVelocity();
 	public SimulationNode getTargetNode();
 
 	public DelayData getDelayData();
