+++ src/test/java/cz/cvut/fel/aic/agentpolis/simmodel/environment/model/congestion/drive/support/DriveAgent.java
@@ -89,7 +89,7 @@ public class DriveAgent extends Agent implements Driver<PhysicalVehicle>{
 
 
 	@Override
-	public double getVelocity() {
+	public int getVelocity() {
 		return 15;
 	}
